# secp256k1: Improve and optimize nonce generation.