--- a/dcrec/secp256k1/schnorr/threshold.go
+++ b/dcrec/secp256k1/schnorr/threshold.go
@@ -56,11 +56,10 @@ func combinePubkeys(pks []*secp256k1.PublicKey) *secp256k1.PublicKey {
 
 // nonceRFC6979 is a local instantiation of deterministic nonce generation
 // by the standards of RFC6979.
-func nonceRFC6979(privkey []byte, hash []byte, extra []byte,
-	version []byte) []byte {
+func nonceRFC6979(privkey []byte, hash []byte, extra []byte, version []byte, extraIterations uint32) []byte {
 	pkD := new(big.Int).SetBytes(privkey)
 	defer pkD.SetInt64(0)
-	bigK := secp256k1.NonceRFC6979(pkD, hash, extra, version)
+	bigK := secp256k1.NonceRFC6979(pkD, hash, extra, version, extraIterations)
 	defer bigK.SetInt64(0)
 	k := bigIntToEncodedBytes(bigK)
 	return k[:]