--- a/dcrec/secp256k1/signature.go
+++ b/dcrec/secp256k1/signature.go
@@ -1,5 +1,5 @@
 // Copyright (c) 2013-2014 The btcsuite developers
-// Copyright (c) 2015-2016 The Decred developers
+// Copyright (c) 2015-2020 The Decred developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
@@ -286,8 +286,7 @@ func hashToInt(hash []byte) *big.Int {
 // of the loop * 2 - on the first iteration of j we do the R case, else the -R
 // case in step 1.6. This counter is used in the Decred compressed signature
 // format and thus we match bitcoind's behaviour here.
-func recoverKeyFromSignature(sig *Signature, msg []byte,
-	iter int, doChecks bool) (*PublicKey, error) {
+func recoverKeyFromSignature(sig *Signature, msg []byte, iter int, doChecks bool) (*PublicKey, error) {
 	// 1.1 x = (n * i) + r
 	curve := S256()
 	Rx := new(big.Int).Mul(curve.Params().N,
@@ -353,8 +352,7 @@ func recoverKeyFromSignature(sig *Signature, msg []byte,
 // returned in the format:
 // <(byte of 27+public key solution)+4 if compressed >< padded bytes for signature R><padded bytes for signature S>
 // where the R and S parameters are padded up to the bitlength of the curve.
-func SignCompact(key *PrivateKey,
-	hash []byte, isCompressedKey bool) ([]byte, error) {
+func SignCompact(key *PrivateKey, hash []byte, isCompressedKey bool) ([]byte, error) {
 	sig, err := key.Sign(hash)
 	if err != nil {
 		return nil, err
@@ -401,8 +399,7 @@ func SignCompact(key *PrivateKey,
 // Koblitz curve in "curve". If the signature matches then the recovered public
 // key will be returned as well as a boolean if the original key was compressed
 // or not, else an error will be returned.
-func RecoverCompact(signature,
-	hash []byte) (*PublicKey, bool, error) {
+func RecoverCompact(signature, hash []byte) (*PublicKey, bool, error) {
 	bitlen := (S256().BitSize + 7) / 8
 	if len(signature) != 1+bitlen*2 {
 		return nil, false, errors.New("invalid compact signature size")