**This requires PR #2042.**

This consists of several commits which improve the code in regards to the signing process by significantly optimizing the deterministic nonce generation by specializing the implementation specifically for secp256k1 and modifying it in such a way that signing can no longer fail in the extremely unlikely event a returned nonce results in an invalid signature.

It has been split into several commits to help ease the review process.

---

The first commit merely reformats the functions to be consistent with the style used in the rest of the code.  It contains no functional changes.

---

The second commit improves the NonceRFC6979 comment which previously indicated it was to generate an ECDSA nonce, but that is not the case.  Instead, it is a method to generate a deterministic nonce based on a private key and message which for general use in signing applications.  As a case in point, it is used with Schnorr signing as well.

---

The third commit adds a benchmark for NonceRFC6979.

---

The fourth commit optimizes the code for generating a deterministic nonce by specializing the implementation for secp256k1 since it no longer accepts a curve argument.

It also implements a specialized version of HMAC-SHA256 in order to avoid the need to constantly allocate new instances due to the changing key.

The following is a before and after comparison of typical nonce generation:
```
benchmark               old ns/op    new ns/op    delta
---------------------------------------------------------
BenchmarkNonceRFC6979   8998         6428         -28.56%

benchmark               old allocs   new allocs   delta
---------------------------------------------------------
BenchmarkNonceRFC6979   46           17           -63.04%

benchmark               old bytes    new bytes    delta
---------------------------------------------------------
BenchmarkNonceRFC6979   3552         976          -72.52%
```
---

Finally, the firth commit modifies the deterministic nonce generation via RFC6979 to accept an additional parameter for the number of extra iterations to perform to ensure a stream of deterministic nonces can be generated if needed.

It then updates the signing for both ECDSA and Schnorr to make use of the new parameter to increment the iteration in the unlikely event a returned nonce results in an invalid signature so that it will always generate a valid signature.

Finally, it removes the error return from `Sign` since it can no longer fail.